refactoring

+++ src/de/htwg/upfaz/backgammon/gui/Constances.java
@@ -6,7 +6,7 @@ final class Constances {
 
     public static final int DEFAULT_X_RESOLUTION = 740;
     public static final int DEFAULT_Y_RESOLUTION = 710;
-    public static final int TIME_TO_SLEEP_IN_MS = 200;
+    public static final long TIME_TO_SLEEP_IN_MS = 200;
     public static final int FIELD_EATEN_WHITE = 25;
     public static final int FIELD_EATEN_BLACK = 24;
     public static final int FIELD_END_WHITE = 27;
